refactor: refatorando widgets da home page
Ajuste na home page

ref: https://trello.com/c/AgbTqLMe/36-tela-home
