+++ lib/app/modules/main/main_module.dart
@@ -1,19 +1,11 @@
-import 'package:dio/dio.dart';
-import 'package:flutter/material.dart';
 import 'package:flutter_modular/flutter_modular.dart';
-import 'package:plant_care/app/core/env/variables.dart';
-import 'package:plant_care/app/core/services/location/location_service.dart';
 import 'package:plant_care/app/modules/main/submodules/home/home_module.dart';
-import 'package:plant_care/app/modules/main/submodules/home/widgets/weather/repositories/weather_repository.dart';
 import 'package:plant_care/app/modules/main/submodules/marketplace/marketplace_module.dart';
 import 'bottom_navigator_store.dart';
 import 'bottom_navigator_widget.dart';
 import 'submodules/community/community_module.dart';
 import 'submodules/configurations/configurations_page.dart';
 import 'submodules/education/education_module.dart';
-import 'submodules/home/repositories/news_repository.dart';
-import 'submodules/home/stores/home_store.dart';
-import 'submodules/home/widgets/weather/page/weather_store.dart';
 
 class MainModule extends Module {
   @override
@@ -31,6 +23,6 @@ class MainModule extends Module {
           ModuleRoute('/education', module: EducationModule()),
           ModuleRoute('/marketplace', module: MarketplaceModule()),
           ChildRoute('/configurations', child: (_, __) => ConfigurationPage()),
-        ]),
+        ],),
   ];
 }
