--- lib/app/modules/main/submodules/home/models/mock_news.dart
@@ -1,2 +1,0 @@
-var MOCK_NEWS =
-    "{\"status\": \"ok\",\"totalResults\": 10,\"articles\": [{\"title\":\"‘Importância do agro para economia deve ser considerada nas negociações do Plano Safra’\",\"author\":\"\",\"source\":{\"id\":\"‘importância-do agro para economia deve ser considerada nas negociações do plano safra’\",\"name\":\"‘Importância do agro para economia deve ser considerada nas negociações do Plano Safra’\"},\"url\":\"https://www.canalrural.com.br/noticias/importancia-do-agro-para-economia-deve-ser-considerada-nas-negociacoes-do-plano-safra/\",\"categoria\":\"RECONHECIMENTO \",\"description\":\"Segundo a gerente técnica da Faemg, o desempenho positivo do agro no primeiro trimestre contribuiu para o desenvolvimento da economia \",\"urlToImage\":\"https://imagens-cdn.canalrural.com.br/2019/01/dinheirosoja03-377x277.jpg\",\"publishedAt\":\"2021-11-06T23:25:00.000Z\",\"content\":\"Segundo a gerente técnica da Faemg, o desempenho positivo do agro no primeiro trimestre contribuiu para o desenvolvimento da economia \"},{\"title\":\"‘Teremos um Plano Safra chocho este ano’, diz especialista em crédito rural\",\"author\":\"\",\"source\":{\"id\":\"‘teremos-um plano safra chocho este ano’, diz especialista em crédito rural\",\"name\":\"‘Teremos um Plano Safra chocho este ano’, diz especialista em crédito rural\"},\"url\":\"https://www.canalrural.com.br/programas/informacao/rural-noticias/teremos-um-plano-safra-chocho-este-ano-diz-especialista-em-credito-rural/\",\"categoria\":\"ENTREVISTA\",\"description\":\"Ademiro Vian avalia que o planto será feito em cima da hora e sem muita margem para inovar; confira a entrevista\",\"urlToImage\":\"https://imagens-cdn.canalrural.com.br/2018/06/dinheirocifrao-320x213.jpg\",\"publishedAt\":\"2021-11-06T23:25:00.000Z\",\"content\":\"Ademiro Vian avalia que o planto será feito em cima da hora e sem muita margem para inovar; confira a entrevista\"},{\"title\":\"Neutralidade climática deve trazer chuvas abaixo da média nos próximos meses\",\"author\":\"\",\"source\":{\"id\":\"neutralidade-climática deve trazer chuvas abaixo da média nos próximos meses\",\"name\":\"Neutralidade climática deve trazer chuvas abaixo da média nos próximos meses\"},\"url\":\"https://www.canalrural.com.br/noticias/neutralidade-climatica-deve-trazer-chuvas-abaixo-da-media-nos-proximos-meses/\",\"categoria\":\"O QUE ESPERAR DO CLIMA?\",\"description\":\"As chuvas ficarão acima do esperado para o centro e Sul do país. No entanto, as precipitações não devem atingir as lavouras de inverno\",\"urlToImage\":\"https://imagens-cdn.canalrural.com.br/2019/03/drops-2404441_960_720-320x217.jpg\",\"publishedAt\":\"2021-11-06T23:20:00.000Z\",\"content\":\"As chuvas ficarão acima do esperado para o centro e Sul do país. No entanto, as precipitações não devem atingir as lavouras de inverno\"},{\"title\":\"Crédito no Plano Safra será suficiente e caberá no bolso do produtor?\",\"author\":\"\",\"source\":{\"id\":\"crédito-no plano safra será suficiente e caberá no bolso do produtor?\",\"name\":\"Crédito no Plano Safra será suficiente e caberá no bolso do produtor?\"},\"url\":\"https://www.canalrural.com.br/programas/informacao/rural-noticias/credito-no-plano-safra-sera-suficiente-e-cabera-no-bolso-do-produtor/\",\"categoria\":\"OPINIÃO\",\"description\":\"Miguel Daoud faz uma análise sobre o que está sendo discutido pelo governo para o novo Plano Safra\",\"urlToImage\":\"https://imagens-cdn.canalrural.com.br/wp-content/uploads/Daoudcenario-640x640.jpg\",\"publishedAt\":\"2021-11-06T23:20:00.000Z\",\"content\":\"Miguel Daoud faz uma análise sobre o que está sendo discutido pelo governo para o novo Plano Safra\"},{\"title\":\"MT: produção de calcário deve ter alta de 12%, mas problemas logísticos prejudicam vendas\",\"author\":\"\",\"source\":{\"id\":\"mt:-produção de calcário deve ter alta de 12%, mas problemas logísticos prejudicam vendas\",\"name\":\"MT: produção de calcário deve ter alta de 12%, mas problemas logísticos prejudicam vendas\"},\"url\":\"https://www.canalrural.com.br/programas/informacao/rural-noticias/mt-producao-de-calcario-deve-ter-alta-de-12-mas-problemas-logisticos-prejudicam-vendas/\",\"categoria\":\"DEMANDA AQUECIDA\",\"description\":\"A procura aquecida movimenta o pátio das indústrias nesta época do ano, que marca o início da chamada safra do calcário\",\"urlToImage\":\"https://imagens-cdn.canalrural.com.br/2018/07/calcario_640x480_acf_cropped.jpg\",\"publishedAt\":\"2021-11-06T23:15:00.000Z\",\"content\":\"A procura aquecida movimenta o pátio das indústrias nesta época do ano, que marca o início da chamada safra do calcário\"},{\"title\":\"Exportações de café do Brasil recuam 20% em maio\",\"author\":\"\",\"source\":{\"id\":\"exportações-de café do brasil recuam 20% em maio\",\"name\":\"Exportações de café do Brasil recuam 20% em maio\"},\"url\":\"https://www.canalrural.com.br/noticias/cafe-exportacoes-brasil-maio/\",\"categoria\":\"RESULTADO NEGATIVO \",\"description\":\"Segundo o Cecafé, a queda nas vendas foi impulsionada por problemas logísticos e mudança no certificado das exportações do grão\",\"urlToImage\":\"https://imagens-cdn.canalrural.com.br/wp-content/uploads/tina-guina-obV_LM0KjxY-unsplash-640x427.jpg\",\"publishedAt\":\"2021-11-06T23:10:00.000Z\",\"content\":\"Segundo o Cecafé, a queda nas vendas foi impulsionada por problemas logísticos e mudança no certificado das exportações do grão\"},{\"title\":\"O que está derrubando as perspectivas para a safra de grãos?\",\"author\":\"\",\"source\":{\"id\":\"o-que está derrubando as perspectivas para a safra de grãos?\",\"name\":\"O que está derrubando as perspectivas para a safra de grãos?\"},\"url\":\"https://www.canalrural.com.br/programas/informacao/rural-noticias/o-que-esta-derrubando-as-perspectivas-para-a-safra-de-graos/\",\"categoria\":\"COMENTÁRIO\",\"description\":\"Celso Grisi explica que o clima é o fator determinante na produção agrícola e na composição de preços\",\"urlToImage\":\"https://imagens-cdn.canalrural.com.br/2019/09/grãos-640x480.jpg\",\"publishedAt\":\"2021-11-06T23:00:00.000Z\",\"content\":\"Celso Grisi explica que o clima é o fator determinante na produção agrícola e na composição de preços\"},{\"title\":\"Arroba do boi gordo registra nova alta e sobe até R\$ 2 nesta sexta, 11\",\"author\":\"\",\"source\":{\"id\":\"arroba-do boi gordo registra nova alta e sobe até r\$ 2 nesta sexta, 11\",\"name\":\"Arroba do boi gordo registra nova alta e sobe até R\$ 2 nesta sexta, 11\"},\"url\":\"https://www.canalrural.com.br/noticias/arroba-do-boi-gordo-registra-nova-alta-e-sobe-ate-r-2-nesta-sexta-11/\",\"categoria\":\"SUBIU MAIS UMA VEZ\",\"description\":\"Segundo o analista da Safras &amp; Mercado, Fernando Henrique Iglesias, os preços estão mais altos em grande parte da região Centro-Oeste\",\"urlToImage\":\"https://imagens-cdn.canalrural.com.br/wp-content/uploads/bois-no-pasto-640x640.jpg\",\"publishedAt\":\"2021-11-06T21:49:00.000Z\",\"content\":\"Segundo o analista da Safras &amp; Mercado, Fernando Henrique Iglesias, os preços estão mais altos em grande parte da região Centro-Oeste\"},{\"title\":\"Governo autoriza retomada de crédito rural do Plano Safra 20/21 com juros subsidiados\",\"author\":\"\",\"source\":{\"id\":\"governo-autoriza retomada de crédito rural do plano safra 20/21 com juros subsidiados\",\"name\":\"Governo autoriza retomada de crédito rural do Plano Safra 20/21 com juros subsidiados\"},\"url\":\"https://www.canalrural.com.br/noticias/autoriza-credito-plano-safra/\",\"categoria\":\"OFÍCIO\",\"description\":\"A retomada acontece no mesmo dia em que o presidente Jair Bolsonaro sancionou o Projeto de Lei do Congresso que recompõe o Orçamento Geral da União\",\"urlToImage\":\"https://imagens-cdn.canalrural.com.br/2018/10/grãos_milho_mercado_dinheiro-7-1-320x213.jpg\",\"publishedAt\":\"2021-11-06T20:50:00.000Z\",\"content\":\"A retomada acontece no mesmo dia em que o presidente Jair Bolsonaro sancionou o Projeto de Lei do Congresso que recompõe o Orçamento Geral da União\"},{\"title\":\"Petrobras reduz preço da gasolina em R\$ 0,05 nas refinarias\",\"author\":\"\",\"source\":{\"id\":\"petrobras-reduz preço da gasolina em r\$ 0,05 nas refinarias\",\"name\":\"Petrobras reduz preço da gasolina em R\$ 0,05 nas refinarias\"},\"url\":\"https://www.canalrural.com.br/agronegocio/petrobras-reduz-preco-gasolina-refinarias/\",\"categoria\":\"A PARTIR DESTE SÁBADO, 12\",\"description\":\"O novo valor para a gasolina passa a vigorar a partir deste sábado, 12, devendo ficar em R\$ 2,53 por litro vendido nas refinarias\",\"urlToImage\":\"https://imagens-cdn.canalrural.com.br/wp-content/uploads/petobras-gasolina-640x640.jpg\",\"publishedAt\":\"2021-11-06T20:40:00.000Z\",\"content\":\"O novo valor para a gasolina passa a vigorar a partir deste sábado, 12, devendo ficar em R\$ 2,53 por litro vendido nas refinarias\"},{\"title\":\"‘Importância do agro para economia deve ser considerada nas negociações do Plano Safra’\",\"author\":\"\",\"source\":{\"id\":\"‘importância-do agro para economia deve ser considerada nas negociações do plano safra’\",\"name\":\"‘Importância do agro para economia deve ser considerada nas negociações do Plano Safra’\"},\"url\":\"https://www.canalrural.com.br/noticias/importancia-do-agro-para-economia-deve-ser-considerada-nas-negociacoes-do-plano-safra/\",\"categoria\":\"RECONHECIMENTO \",\"description\":\"\",\"urlToImage\":\"https://imagens-cdn.canalrural.com.br/2019/01/dinheirosoja03-377x277.jpg\",\"publishedAt\":\"2021-11-06T23:25:00.000Z\",\"content\":\"\"},{\"title\":\"‘Teremos um Plano Safra chocho este ano’, diz especialista em crédito rural\",\"author\":\"\",\"source\":{\"id\":\"‘teremos-um plano safra chocho este ano’, diz especialista em crédito rural\",\"name\":\"‘Teremos um Plano Safra chocho este ano’, diz especialista em crédito rural\"},\"url\":\"https://www.canalrural.com.br/programas/informacao/rural-noticias/teremos-um-plano-safra-chocho-este-ano-diz-especialista-em-credito-rural/\",\"categoria\":\"ENTREVISTA\",\"description\":\"\",\"urlToImage\":\"https://imagens-cdn.canalrural.com.br/2018/06/dinheirocifrao-320x213.jpg\",\"publishedAt\":\"2021-11-06T23:25:00.000Z\",\"content\":\"\"},{\"title\":\"Neutralidade climática deve trazer chuvas abaixo da média nos próximos meses\",\"author\":\"\",\"source\":{\"id\":\"neutralidade-climática deve trazer chuvas abaixo da média nos próximos meses\",\"name\":\"Neutralidade climática deve trazer chuvas abaixo da média nos próximos meses\"},\"url\":\"https://www.canalrural.com.br/noticias/neutralidade-climatica-deve-trazer-chuvas-abaixo-da-media-nos-proximos-meses/\",\"categoria\":\"O QUE ESPERAR DO CLIMA?\",\"description\":\"\",\"urlToImage\":\"https://imagens-cdn.canalrural.com.br/2019/03/drops-2404441_960_720-320x217.jpg\",\"publishedAt\":\"2021-11-06T23:20:00.000Z\",\"content\":\"\"},{\"title\":\"Crédito no Plano Safra será suficiente e caberá no bolso do produtor?\",\"author\":\"\",\"source\":{\"id\":\"crédito-no plano safra será suficiente e caberá no bolso do produtor?\",\"name\":\"Crédito no Plano Safra será suficiente e caberá no bolso do produtor?\"},\"url\":\"https://www.canalrural.com.br/programas/informacao/rural-noticias/credito-no-plano-safra-sera-suficiente-e-cabera-no-bolso-do-produtor/\",\"categoria\":\"OPINIÃO\",\"description\":\"\",\"urlToImage\":\"https://imagens-cdn.canalrural.com.br/wp-content/uploads/Daoudcenario-640x640.jpg\",\"publishedAt\":\"2021-11-06T23:20:00.000Z\",\"content\":\"\"},{\"title\":\"MT: produção de calcário deve ter alta de 12%, mas problemas logísticos prejudicam vendas\",\"author\":\"\",\"source\":{\"id\":\"mt:-produção de calcário deve ter alta de 12%, mas problemas logísticos prejudicam vendas\",\"name\":\"MT: produção de calcário deve ter alta de 12%, mas problemas logísticos prejudicam vendas\"},\"url\":\"https://www.canalrural.com.br/programas/informacao/rural-noticias/mt-producao-de-calcario-deve-ter-alta-de-12-mas-problemas-logisticos-prejudicam-vendas/\",\"categoria\":\"DEMANDA AQUECIDA\",\"description\":\"\",\"urlToImage\":\"https://imagens-cdn.canalrural.com.br/2018/07/calcario_640x480_acf_cropped.jpg\",\"publishedAt\":\"2021-11-06T23:15:00.000Z\",\"content\":\"\"},{\"title\":\"Exportações de café do Brasil recuam 20% em maio\",\"author\":\"\",\"source\":{\"id\":\"exportações-de café do brasil recuam 20% em maio\",\"name\":\"Exportações de café do Brasil recuam 20% em maio\"},\"url\":\"https://www.canalrural.com.br/noticias/cafe-exportacoes-brasil-maio/\",\"categoria\":\"RESULTADO NEGATIVO \",\"description\":\"\",\"urlToImage\":\"https://imagens-cdn.canalrural.com.br/wp-content/uploads/tina-guina-obV_LM0KjxY-unsplash-640x427.jpg\",\"publishedAt\":\"2021-11-06T23:10:00.000Z\",\"content\":\"\"},{\"title\":\"O que está derrubando as perspectivas para a safra de grãos?\",\"author\":\"\",\"source\":{\"id\":\"o-que está derrubando as perspectivas para a safra de grãos?\",\"name\":\"O que está derrubando as perspectivas para a safra de grãos?\"},\"url\":\"https://www.canalrural.com.br/programas/informacao/rural-noticias/o-que-esta-derrubando-as-perspectivas-para-a-safra-de-graos/\",\"categoria\":\"COMENTÁRIO\",\"description\":\"\",\"urlToImage\":\"https://imagens-cdn.canalrural.com.br/2019/09/grãos-640x480.jpg\",\"publishedAt\":\"2021-11-06T23:00:00.000Z\",\"content\":\"\"},{\"title\":\"Arroba do boi gordo registra nova alta e sobe até R\$ 2 nesta sexta, 11\",\"author\":\"\",\"source\":{\"id\":\"arroba-do boi gordo registra nova alta e sobe até r\$ 2 nesta sexta, 11\",\"name\":\"Arroba do boi gordo registra nova alta e sobe até R\$ 2 nesta sexta, 11\"},\"url\":\"https://www.canalrural.com.br/noticias/arroba-do-boi-gordo-registra-nova-alta-e-sobe-ate-r-2-nesta-sexta-11/\",\"categoria\":\"SUBIU MAIS UMA VEZ\",\"description\":\"\",\"urlToImage\":\"https://imagens-cdn.canalrural.com.br/wp-content/uploads/bois-no-pasto-640x640.jpg\",\"publishedAt\":\"2021-11-06T21:49:00.000Z\",\"content\":\"\"},{\"title\":\"Governo autoriza retomada de crédito rural do Plano Safra 20/21 com juros subsidiados\",\"author\":\"\",\"source\":{\"id\":\"governo-autoriza retomada de crédito rural do plano safra 20/21 com juros subsidiados\",\"name\":\"Governo autoriza retomada de crédito rural do Plano Safra 20/21 com juros subsidiados\"},\"url\":\"https://www.canalrural.com.br/noticias/autoriza-credito-plano-safra/\",\"categoria\":\"OFÍCIO\",\"description\":\"\",\"urlToImage\":\"https://imagens-cdn.canalrural.com.br/2018/10/grãos_milho_mercado_dinheiro-7-1-320x213.jpg\",\"publishedAt\":\"2021-11-06T20:50:00.000Z\",\"content\":\"\"},{\"title\":\"Petrobras reduz preço da gasolina em R\$ 0,05 nas refinarias\",\"author\":\"\",\"source\":{\"id\":\"petrobras-reduz preço da gasolina em r\$ 0,05 nas refinarias\",\"name\":\"Petrobras reduz preço da gasolina em R\$ 0,05 nas refinarias\"},\"url\":\"https://www.canalrural.com.br/agronegocio/petrobras-reduz-preco-gasolina-refinarias/\",\"categoria\":\"A PARTIR DESTE SÁBADO, 12\",\"description\":\"\",\"urlToImage\":\"https://imagens-cdn.canalrural.com.br/wp-content/uploads/petobras-gasolina-640x640.jpg\",\"publishedAt\":\"2021-11-06T20:40:00.000Z\",\"content\":\"\"}]}";
