+++ lib/app/modules/main/submodules/home/widgets/news/models/news.dart
@@ -1,4 +1,4 @@
-import '../../../../../core/models/source.dart';
+import 'source.dart';
 
 class NewsModel {
   final Source? source;
