+++ lib/app/modules/main/submodules/home/widgets/news/repositories/news_repository.dart
@@ -1,9 +1,6 @@
-import 'dart:convert';
-
 import 'package:dio/dio.dart';
 import 'package:flutter_modular/flutter_modular.dart';
-import 'package:plant_care/app/modules/main/submodules/home/models/mock_news.dart';
-import 'package:plant_care/app/modules/main/submodules/home/models/news.dart';
+import 'package:plant_care/app/modules/main/submodules/home/widgets/news/models/news.dart';
 
 class NewsRepository extends Disposable {
   final Dio _http;
@@ -11,18 +8,6 @@ class NewsRepository extends Disposable {
   NewsRepository(this._http);
 
   Future<List<NewsModel>> getAll() async {
-    // Response response = await _http.get('/everything', queryParameters: {
-    //   'apiKey': '2eb849c122b546d3929678e27a902aa6',
-    //   'q': 'agro'
-    // });
-
-    // if (response.statusCode != 200) {
-    //   return <NewsModel>[];
-    // }
-
-    // var data = response.data['articles'];
-    // var response = jsonDecode(MOCK_NEWS);
-
     Response response = await _http.get('/news',
         options: Options(headers: {
           'Origin': 'http://localhost',
@@ -37,6 +22,5 @@ class NewsRepository extends Disposable {
   }
 
   @override
-  void dispose() {
-  }
+  void dispose() {}
 }
